requirements.txt
Django==4.2.20
djangorestframework==3.16.0
drf-yasg==1.21.10
django-filter==25.1
django-timezone-field==7.1
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
psycopg2-binary==2.9.10
python-decouple==3.8
googlemaps==4.10.0
requests==2.32.3
pytest==8.3.5
pytest-django==4.11.1
factory_boy==3.3.3
Faker==37.1.0
